falta por fijar

+++ u25-Ejercicio1/src/main/java/com/team02/u25/ejercicio1/dao/IArticuloDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.team02.u25.ejercicio1.dto.Articulo;
 
-public interface IArticuloDAO extends JpaRepository<Articulo, Long>{
+public interface IArticuloDAO extends JpaRepository<Articulo, Integer>{
 
 }
